pages de contact et mentions légales, style

+++ src/Entity/RecipeIngredient.php
@@ -49,12 +49,12 @@ class RecipeIngredient
         return $this->id;
     }
 
-    public function getQuantity(): ?int
+    public function getQuantity(): ?float
     {
         return $this->quantity;
     }
 
-    public function setQuantity(int $quantity): self
+    public function setQuantity(float $quantity): self
     {
         $this->quantity = $quantity;
